+++ retroauth/src/main/java/com/andretietz/retroauth/Retroauth.kt
@@ -31,7 +31,7 @@ import java.util.concurrent.Executor
 class Retroauth private constructor() {
 
     class Builder<out OWNER_TYPE : Any, OWNER : Any, TOKEN_TYPE : Any, TOKEN : Any> @JvmOverloads constructor(
-            private val tokenProvider: TokenProvider<OWNER_TYPE, OWNER, TOKEN_TYPE, TOKEN>,
+            private val authenticator: Authenticator<OWNER_TYPE, OWNER, TOKEN_TYPE, TOKEN>,
             private val ownerManager: OwnerManager<OWNER_TYPE, OWNER, TOKEN_TYPE>,
             private val tokenStorage: TokenStorage<OWNER, TOKEN_TYPE, TOKEN>,
             private val methodCache: MethodCache<OWNER_TYPE, TOKEN_TYPE> = MethodCache.DefaultMethodCache()
@@ -116,7 +116,7 @@ class Retroauth private constructor() {
 
             // creating a custom calladapter to handle authentication
             val callAdapter = RetroauthCallAdapterFactory(callAdapterFactories,
-                    tokenProvider,
+                    authenticator,
                     methodCache)
 
             // use this callAdapter to create the retrofit object
@@ -126,7 +126,7 @@ class Retroauth private constructor() {
 
             // create the okhttp interceptor to intercept requests
             val interceptor = CredentialInterceptor(
-                    tokenProvider,
+                    authenticator,
                     ownerManager,
                     tokenStorage,
                     methodCache