+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * Better separation of OwnerManager and TokenManager
 * Fixed Demo app which is using Facebook with [scribe](https://github.com/scribejava/scribejava)
 * Added functionalities to refresh the token before the authenticated call is called.
+* Renaming TokenProvider to Authenticator
 
 ## 2.3.1 (2018-01-15)
 * Updating release script